requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0

# Azure authentication
azure-identity>=1.15.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For enhanced logging
structlog>=23.1.0
